MessageSource를 이용한 Exception 처리

+++ server/src/com/riding/springboot/user/service/security/CustomUserDetailService.java
@@ -14,6 +14,6 @@ public class CustomUserDetailService implements UserDetailsService {
     private final UserRepository userJpaRepo;
 
     public UserDetails loadUserByUsername(String userPk) {
-        return userJpaRepo.findById(Integer.valueOf(userPk)).orElseThrow(CUserNotFoundException::new);
+        return userJpaRepo.findById(Long.valueOf(userPk)).orElseThrow(CUserNotFoundException::new);
     }
 }
